refactor: remove react-native dependency from create function
